+++ presentacion/modelos_qt/__init__.py
@@ -1,0 +1,1 @@
+"""Modelos Qt de la capa de presentación."""